A

+++ src/main/java/com/uc10e3/model/Podcast.java
@@ -18,7 +18,7 @@ import jakarta.persistence.Id;
 public class Podcast {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
     private String produtor;
     private String nomeDoEpisodio;
     private int numeroDoEpisodio;
@@ -51,11 +51,11 @@ public class Podcast {
      * @return ID numérico do podcast.
      */
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
